+++ src/_pytask/mark/__init__.py
@@ -133,9 +133,7 @@ class KeywordMatcher:
         mapped_names = {task.name}
 
         # Add the names attached to the current function through direct assignment.
-        function_obj = task.function
-        if function_obj:
-            mapped_names.update(function_obj.__dict__)
+        mapped_names.update(task.function.__dict__)
 
         # Add the markers to the keywords as we no longer handle them correctly.
         mapped_names.update(mark.name for mark in task.markers)
@@ -149,7 +147,7 @@ class KeywordMatcher:
         return any(subname in name for name in names)
 
 
-def select_by_keyword(session: Session, dag: nx.DiGraph) -> set[str]:
+def select_by_keyword(session: Session, dag: nx.DiGraph) -> set[str] | None:
     """Deselect tests by keywords."""
     keywordexpr = session.config["expression"]
     if not keywordexpr:
@@ -204,7 +202,7 @@ class MarkMatcher:
         return name in self.own_mark_names
 
 
-def select_by_mark(session: Session, dag: nx.DiGraph) -> set[str]:
+def select_by_mark(session: Session, dag: nx.DiGraph) -> set[str] | None:
     """Deselect tests by marks."""
     matchexpr = session.config["marker_expression"]
     if not matchexpr: